requirements.txt
alembic==1.3.0
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
bs4==0.0.1
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
google-api-python-client==1.7.11
google-auth==1.8.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
Mako==1.1.0
MarkupSafe==1.1.1
mechanize==0.4.4
more-itertools==7.2.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyJWT==1.7.1
pymemcache==2.2.2
pyrsistent==0.15.5
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.10
uritemplate==3.0.0
urllib3==1.25.7
webencodings==0.5.1
Werkzeug==0.16.0
zipp==0.6.0
